Update mpi_quad.c
The MPI_Bcast function call specified type MPI_INT to describe the variable "n" which had been initialized as "long int". This causes incorrect results for the integer comparison which controls the for loop when the code is compiled with recent (at least 2020 - 2022) Intel MPI compilers. Correct typing is preferred.